requirements.txt
mesop
google-generativeai
ollama
groq
llama-index
llama-index-llms-anthropic
langchain_community
langchain